oop programming

+++ programming_paradigm/main.py
@@ -6,8 +6,8 @@ def main():
         print("Usage: python main.py <numerator> <denominator>")
         sys.exit(1)
     try:
-        numerator = int(sys.argv[1])
-        denominator = int(sys.argv[2])
+        numerator = float(sys.argv[1])
+        denominator = float(sys.argv[2])
         result = safe_divide(numerator, denominator)
         print(result)
     except ValueError:
